Move further services to core
