+++ org.opentosca.bus/org.opentosca.bus.management.service/src/main/java/org/opentosca/bus/management/service/impl/PluginRegistry.java
@@ -6,6 +6,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import javax.inject.Inject;
+import javax.inject.Named;
 import javax.inject.Singleton;
 
 import org.opentosca.bus.management.deployment.plugin.IManagementBusDeploymentPluginService;
@@ -15,6 +16,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 @Singleton
+@Named("managementBusPluginRegistry")
 public class PluginRegistry {
 
     private final Map<String, IManagementBusInvocationPluginService> invocationPluginServices =
