+++ org.opentosca.container.core/src/main/java/org/opentosca/container/core/next/services/SituationInstanceService.java
@@ -1,4 +1,4 @@
-package org.opentosca.container.api.service;
+package org.opentosca.container.core.next.services;
 
 import java.util.Collection;
 import java.util.Map;
